+++ CHANGELOG.md
@@ -14,6 +14,10 @@ the [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format.
 - Added the optional `preserve_ids` parameter to `OpenCC` for preserving characters inside Unicode IDS structures during
   conversion.
 
+### Changed
+
+- CLI: Optimized subcommand `office`
+
 ---
 
 ## [0.10.0] - 2026-06-17